refactor(studio): reuse selected-branch ordering in token estimator
