+++ studio/frontend/src/features/chat/utils/refresh-context-usage.ts
@@ -15,6 +15,9 @@ import { isExternalModelId } from "../external-providers";
 import { useChatRuntimeStore } from "../stores/chat-runtime-store";
 import type { MessageRecord } from "../types";
 import { listStoredChatMessages } from "./chat-history-storage";
+import { orderBySelectedBranch } from "./order-selected-branch";
+
+export { orderBySelectedBranch };
 
 // Per thread, not per module: in compare mode a hidden pane's history load would otherwise
 // invalidate the visible thread's count, blanking the bar.
@@ -78,45 +81,6 @@ function storedMessageToRunMessage(record: MessageRecord): ThreadMessage {
       unstable_state: null,
     },
   };
-}
-
-// Same order the history adapter sorts stored records in before importing them.
-const ROLE_ORDER: Record<string, number> = { system: 0, user: 1, assistant: 2 };
-
-/**
- * The displayed branch, rebuilt as the history adapter does: sort by (createdAt, role, id), parent
- * legacy records to the previous one, then walk the last record's ancestor chain. Greedy
- * newest-child descent would pick another branch and drop pre-parentId history.
- */
-function orderBySelectedBranch<T extends MessageRecord>(messages: T[]): T[] {
-  const sorted = messages.slice().sort((a, b) => {
-    if (a.createdAt !== b.createdAt) return a.createdAt - b.createdAt;
-    const aOrder = ROLE_ORDER[a.role] ?? 99;
-    const bOrder = ROLE_ORDER[b.role] ?? 99;
-    if (aOrder !== bOrder) return aOrder - bOrder;
-    return a.id < b.id ? -1 : a.id > b.id ? 1 : 0;
-  });
-
-  const byId = new Map<string, T>();
-  const parentOf = new Map<string, string | null>();
-  let previousId: string | null = null;
-  for (const m of sorted) {
-    byId.set(m.id, m);
-    parentOf.set(m.id, m.parentId ?? previousId);
-    previousId = m.id;
-  }
-
-  const chain: T[] = [];
-  const seen = new Set<string>();
-  let cur: string | null = sorted.at(-1)?.id ?? null;
-  while (cur != null && !seen.has(cur)) {
-    seen.add(cur);
-    const record = byId.get(cur);
-    if (!record) break;
-    chain.push(record);
-    cur = parentOf.get(cur) ?? null;
-  }
-  return chain.reverse();
 }
 
 /** Rolling 32-bit hash. Only has to change when the input does, not resist an adversary. */
